--- app/controllers/tops_controller.rb
@@ -1,4 +1,0 @@
-class TopsController < ApplicationController
-  layout 'top_layout'
-  def top; end
-end
