--- app/views/layouts/top_layout.html.erb
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title><%= full_title(yield(:title))%></title>
-    <%= csrf_meta_tags %>
-    <%= csp_meta_tag %>
-    <%= stylesheet_link_tag 'application', media: 'all', 'data-turbolinks-track': 'reload' %>
-    <%= javascript_include_tag 'application', 'data-turbolinks-track': 'reload' %>
-    <%= render 'layouts/shim' %>
-  </head>
-  <body class="overflow-hidden">
-    <main class="container-fuild">
-      <% flash.each do |message_type, message| %>
-        <%= content_tag(:div, message, class: "alert alert-#{message_type}") %>
-      <% end %>
-      <%= yield %>
-    </main>
-  </body>
-</html>
