+++ app/views/static_pages/top.html.erb
