+++ config/routes.rb
@@ -3,7 +3,7 @@
 Rails.application.routes.draw do
   get 'password_resets/new'
   get 'password_resets/edit'
-  root 'tops#top'
+  root 'static_pages#top'
   get  '/signup', to: 'users#new'
   post '/signup', to: 'users#create'
   get  '/login',   to: 'sessions#new'
@@ -14,7 +14,7 @@ Rails.application.routes.draw do
   get '/about', to: 'static_pages#about'
   get 'songs/search', to: 'songs#search'
   post '/search_result', to: 'static_pages#search_result'
-  mount LetterOpenerWeb::Engine, at: "/letter_opener" if Rails.env.development?
+  mount LetterOpenerWeb::Engine, at: '/letter_opener' if Rails.env.development?
   resources :users do
     member do
       get :following, :followers
