+++ spec/controllers/static_pages_controller_spec.rb
@@ -1,6 +1,20 @@
 require 'rails_helper'
 
 RSpec.describe StaticPagesController, type: :controller do
+  describe '#top' do
+    before do
+      get :top
+    end
+
+    it '正常にレスポンスを返すこと' do
+      expect(response).to be_successful
+    end
+
+    it 'ステータスが200であること' do
+      expect(response.status).to eq 200
+    end
+  end
+
   describe '#home' do
     before do
       get :home
